Introduce datawire/envconfig and move environment parsing to one place
The `datawire/envconfig` replaces sethvargo/go-envconfig in this commit,
enabling us to unify all environment string parsing to one place. This
commit also changes the way defaults are handled so that all such values
are declared in the Helm chart, and only in the Helm chart. The
rationale for this is that some values are used in more places than the
environment, and would therefore need to have defaults in more than one
place unless they are in the Helm chart values.

+++ cmd/traffic/cmd/manager/service.go
@@ -158,7 +158,7 @@ func (m *Manager) GetCloudConfig(ctx context.Context, _ *empty.Empty) (*rpc.Amba
 // GetTelepresenceAPI returns information about the TelepresenceAPI server
 func (m *Manager) GetTelepresenceAPI(ctx context.Context, e *empty.Empty) (*rpc.TelepresenceAPIInfo, error) {
 	env := managerutil.GetEnv(ctx)
-	return &rpc.TelepresenceAPIInfo{Port: env.APIPort}, nil
+	return &rpc.TelepresenceAPIInfo{Port: int32(env.APIPort)}, nil
 }
 
 // ArriveAsClient establishes a session between a client and the Manager.
